ENH: add true Buttorworth variant
The current Butterworth filter is actually the square of a Butterworth filter.
Add a true_butterworth keyword argument defaulting to False to allow use of the standard Butterworth.
Some textbooks use the squared form, but this is not how a Butterworth is defined in signal processing.

Added test cases verifying the expected amplitudes at the cutoff in each case.